refactor: move BundleChunker to client, SizeExtensions to extension package
