+++ src/test/kotlin/io/github/zenhelix/gradle/plugin/extension/SizeExtensionsTest.kt
@@ -1,4 +1,4 @@
-package io.github.zenhelix.gradle.plugin.utils
+package io.github.zenhelix.gradle.plugin.extension
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
